Refactor: Change directory structure to suit Main content tabs
