+++ src/components/Main/Content/Overview/MainContent/MainContent.vue
@@ -1,0 +1,19 @@
+<template>
+      <!-- Main Content and Buttons -->
+      <content-section-two />
+</template>
+
+<script>
+    // Import our components from component directory
+    import SectionTwo from './SectionRight/MainSectionTwo.vue'
+
+    // Use our components in our App.vue template
+    export default {
+        components: {
+            contentSectionTwo: SectionTwo
+        }
+    }
+</script>
+
+<style>
+</style>
