+++ src/components/Main/Content/Overview/MainContent/SectionRight/Content/EuropeContent/Graph/GraphContent.vue
