+++ src/components/Main/Content/Overview/MainContent/SectionRight/Content/EuropeContent/GraphFilter.vue
