+++ src/components/Main/Content/Overview/MainContent/SectionRight/Content/SectionTabs.vue
