+++ src/components/Main/SideBar/SectionLeft/SectionOne.vue
@@ -27,9 +27,9 @@
 
 
 <style scoped>
-    @import './../../../Assets/normalize.css';
-    @import './../../../Assets/webflow.css';
-    @import './../../../Assets/custom.css';
+    @import './../../../../Assets/normalize.css';
+    @import './../../../../Assets/webflow.css';
+    @import './../../../../Assets/custom.css';
 
     /* Border Style Pattern */
     .section-front-box {
